refactor: transcript cjsToEsm to ES6
